The 'seqList' passed into each function in this library-package is allowed to be a mixed list of integers and strings. The strings are checked for validity, valid being either "A-B", "A-BxN" or just an "A" (where A, B, N are ints). Any strings that are invalid are added into the "nonSeqList" and passed back to the caller once the function finishes processing the entire 'seqList'.
This commit expands what is allowed to be passed as a string entry, in that if any
string entries contain whitespace, then that entry is split into multiple entries,
each of which is processed as before.

Note that in the last release, whitespace was totally stripped out of such strings in
expandseq(), which led to possibly undesirable results, for example "20 21" would get
turned into "2021", which is probably not what was wanted.

Now, commas are ALSO allowed in strings and will be treated as whitespace, then
any string entries will be split along whitespace and then processed as before.

Note that "A-B" or "A-BxN" MUST NOT contain whitespace, since "A -B" is invalid
syntax for a range from A to B, and will now be interpreted as just two ints,
namely "A" and "-B".